Update all tooling
