--- .eslintrc
@@ -1,25 +1,0 @@
-{
-  "extends": "airbnb",
-
-  "rules": {
-    "indent": [2, 2],
-    "max-len": [1, 120, 4, {"ignoreUrls": true}],
-    "id-length": [1, {"min": 2, "exceptions": ["x", "y", "e", "i", "j", "k", "d", "n", "_", "$"]}],
-    "object-shorthand": [2, "methods"],
-    "arrow-body-style": [0],
-    "no-new": [1],
-    "no-multi-spaces": [0],
-    "no-warning-comments": [1, { "terms": ["todo", "fixme", "xxx"], "location": "start" }],
-    "jsx-quotes": [2, "prefer-double"],
-    "react/prefer-es6-class": [0],
-    "react/jsx-no-bind": [1],
-    "react/jsx-filename-extension": [0],
-    "react/no-find-dom-node": [0],
-    "import/no-extraneous-dependencies": [0]
-  },
-
-  "env": {
-      "mocha": true,
-      "browser": true
-  }
-}
