+++ .npmignore
@@ -1,15 +1,19 @@
-**/.*
-__tests__
+coverage
+server
+.babelrc
 examples
-
-.editorconfig
-.gitattributes
-.gitignore
-.eslintrc
-.eslintignore
+docs
+lib
 .npmignore
-.travis.yml
+webpack
 .nvmrc
-
-CHANGELOG.md
-server.js
+.travis.yml
+.eslintignore
+.eslintrc
+.editorconfig
+.githooks
+.github
+bin
+.env
+pages
+webpack-stats.json
