+++ package.json
@@ -20,20 +20,88 @@
     "optgroups",
     "multiselect"
   ],
+  "eslintConfig": {
+    "extends": "plugin:springload/recommended",
+    "rules": {
+      "indent": [
+        "error",
+        2
+      ],
+      "react/jsx-indent": [
+        "error",
+        2
+      ],
+      "react/jsx-indent-props": [
+        "error",
+        2
+      ]
+    }
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/dist/",
+      "/pages/",
+      "/examples/test.js"
+    ],
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ]
+  },
+  "babel": {
+    "presets": [
+      [
+        "env",
+        {
+          "targets": {
+            "browsers": [
+              "> 1%",
+              "IE 11"
+            ]
+          },
+          "modules": false
+        }
+      ],
+      "react"
+    ],
+    "env": {
+      "test": {
+        "presets": [
+          [
+            "env",
+            {
+              "targets": {
+                "browsers": [
+                  "> 1%",
+                  "IE 11"
+                ]
+              }
+            }
+          ],
+          "react"
+        ]
+      }
+    }
+  },
   "dependencies": {},
   "devDependencies": {
-    "chai": "^3.5.0",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb": "^10.0.0",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^6.0.0",
-    "mocha": "^3.0.2",
-    "prop-types": "^15.5.10",
-    "rackt-cli": "0.5.4",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "babel-cli": "^6.22.2",
+    "babel-jest": "^20.0.3",
+    "babel-loader": "^7.0.0",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-react": "^6.16.0",
+    "coveralls": "^2.11.15",
+    "enzyme": "^2.8.2",
+    "enzyme-to-json": "^1.5.1",
+    "eslint": "^3.19.0",
+    "eslint-plugin-springload": "^1.0.2",
+    "jest": "^20.0.4",
+    "prop-types": "^15.5.8",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-test-renderer": "^15.5.4",
+    "webpack": "^2.4.1",
+    "webpack-dev-server": "^2.4.5"
   },
   "peerDependencies": {
     "prop-types": "^15.5.8 || ^16.0.0",
@@ -41,11 +109,21 @@
     "react-dom": "^0.14.9 || ^15.3.0 || ^16.0.0"
   },
   "scripts": {
-    "test": "export RACKT_PATH=./node_modules/rackt-cli; ./node_modules/rackt-cli/node_modules/.bin/karma start ./node_modules/rackt-cli/karma.conf.js --single-run --browsers Firefox",
-    "start": "rackt server",
-    "build": "rackt build",
+    "start": "webpack-dev-server --config=webpack/webpack.config.dev.js",
+    "js": "babel -d dist lib",
+    "dist": "npm run js -s",
     "linter:js": "eslint",
-    "lint:js": "npm run linter:js -s -- .",
-    "lint": "npm run lint:js -s"
-  }
+    "lint": "npm run linter:js -s -- .",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
+    "test:watch": "jest --watch",
+    "test:watch:coverage": "jest --watch --coverage",
+    "report:coverage": "open coverage/lcov-report/index.html",
+    "report:build": "NODE_ENV=production webpack --config webpack/webpack.config.prod.js --profile --json > webpack-stats.json",
+    "test:ci": "npm run lint -s && npm run test:coverage -s && npm run dist -s && npm run pages -s",
+    "coveralls": "cat ./coverage/lcov.info | coveralls",
+    "pages": "rm -rf pages && cp -R examples pages && webpack --config=webpack/webpack.config.prod.js",
+    "deploy": "./bin/deploy.sh"
+  },
+  "false": {}
 }
